Use ssize_t for recv etc. return type
Just another 32b/64b compatibility issue

+++ tools/compilers/gcc-glibc/glibc-2.14.1-ros/sysdeps/ros/recv.c
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <sys/socket.h>
 #include <ros/syscall.h>
-int __recv(int s, void *buf, size_t len, int flags) {
+ssize_t __recv(int s, void *buf, size_t len, int flags) {
 	return ros_syscall(SYS_recv, s, buf, len, flags, 0, 0);
 }
